Default workflow runtime kind to claude_code
WORKFLOW.md's runtime_kind/runtime_profile are independent from
[agents] default_agent in the server config — repos without a
project-level WORKFLOW.md were forced to codex_jsonrpc/codex-default
even when the operator had set default_agent = "claude".

Switch the compiled defaults to claude_code/claude-default so all
repos inherit the Claude Code runtime by default. Repos that
genuinely need codex can still override via their own WORKFLOW.md.

Move the workflow.rs test module to workflow_tests.rs via
#[path] so the file stays under the U-16 800-line ceiling and
future default-value tweaks are not blocked by the size guard.
Update default-case asserts to match the new defaults; tests for
explicit-WORKFLOW.md parsing are unchanged.

Also flip harness's own WORKFLOW.md from codex_jsonrpc to
claude_code so the harness repository's intake routes through
claude consistently with the new defaults.
